+++ docker-compose.shared.env.yml
@@ -1,6 +1,8 @@
 version: '3.7'
 services:
   nereid:
+    environment:
+      - NEREID_ASYNC_MODE=replace
     env_file:
       - env.env
   nereid-tests:
