+++ nereid/.coveragerc
@@ -6,4 +6,4 @@ branch = True
 [report]
 include = nereid/*.py
 ignore_errors = True
-omit = nereid/data/*
+omit = nereid/data/*,*main.py
