+++ nereid/nereid/api/api_v1/endpoints_async/reference_data.py
@@ -6,16 +6,13 @@ from typing import Any, Dict, Optional, Union
 from fastapi import APIRouter, Depends, HTTPException
 from fastapi.requests import Request
 from fastapi.responses import FileResponse, ORJSONResponse
-from fastapi.templating import Jinja2Templates
 
 from nereid.api.api_v1.models.reference_models import ReferenceDataResponse
-from nereid.api.api_v1.utils import get_valid_context
+from nereid.api.api_v1.utils import get_valid_context, templates
 from nereid.core.io import load_file, load_json, load_ref_data
 from nereid.src.nomograph.nomo import load_nomograph_mapping
 
 router = APIRouter()
-
-templates = Jinja2Templates(directory="nereid/api/templates")
 
 
 @router.get("/reference_data_file", tags=["reference_data"])
