+++ nereid/nereid/api/api_v1/endpoints_async/watershed.py
@@ -4,11 +4,12 @@ from fastapi import APIRouter, Depends
 from fastapi.responses import ORJSONResponse
 
 import nereid.bg_worker as bg
+from nereid.api.api_v1.async_utils import run_task, standard_json_response
 from nereid.api.api_v1.models.treatment_facility_models import (
     validate_treatment_facility_models,
 )
 from nereid.api.api_v1.models.watershed_models import Watershed, WatershedResponse
-from nereid.api.api_v1.utils import get_valid_context, run_task, standard_json_response
+from nereid.api.api_v1.utils import get_valid_context
 
 router = APIRouter()
