+++ nereid/nereid/api/api_v1/endpoints_sync/watershed.py
@@ -1,6 +1,6 @@
 from typing import Any, Dict, Tuple
 
-from fastapi import APIRouter, Depends, Request
+from fastapi import APIRouter, Depends
 from fastapi.responses import ORJSONResponse
 
 from nereid.api.api_v1.models.treatment_facility_models import (
@@ -8,12 +8,14 @@ from nereid.api.api_v1.models.treatment_facility_models import (
 )
 from nereid.api.api_v1.models.watershed_models import Watershed, WatershedResponse
 from nereid.api.api_v1.utils import get_valid_context
+from nereid.src import tasks
 
 router = APIRouter()
 
 
 def validate_watershed_request(
-    watershed_req: Watershed, context: dict = Depends(get_valid_context),
+    watershed_req: Watershed,
+    context: dict = Depends(get_valid_context),
 ) -> Tuple[Dict[str, Any], Dict[str, Any]]:
 
     watershed: Dict[str, Any] = watershed_req.dict(by_alias=True)
@@ -36,14 +38,13 @@ def validate_watershed_request(
     response_class=ORJSONResponse,
 )
 async def post_solve_watershed(
-    request: Request,
     watershed_pkg: Tuple[Dict[str, Any], Dict[str, Any]] = Depends(
         validate_watershed_request
     ),
 ) -> Dict[str, Any]:
     watershed, context = watershed_pkg
 
-    data = request.app.tasks.solve_watershed(
+    data = tasks.solve_watershed(
         watershed=watershed, treatment_pre_validated=True, context=context
     )
     return {"data": data}
