+++ nereid/nereid/core/config.py
@@ -1,21 +1,20 @@
 import importlib.resources as pkg_resources
-from pathlib import Path
 from typing import Any, Dict, List, Literal, Optional, Union
 
 from pydantic import AnyHttpUrl, BaseSettings, validator
 
 import nereid
 from nereid.core.io import load_cfg
+from nereid.core.utils import get_nereid_path
+
+nereid_path = get_nereid_path()
 
 
 class Settings(BaseSettings):
 
-    with pkg_resources.path("nereid", "__init__.py") as file:
-        _nereid_path = file.parent
-
     API_V1_STR: str = "/api/v1"
     API_LATEST: str = API_V1_STR
-    APP_CONTEXT: Dict[str, Any] = load_cfg(_nereid_path / "core" / "base_config.yml")
+    APP_CONTEXT: Dict[str, Any] = load_cfg(nereid_path / "core" / "base_config.yml")
     APP_CONTEXT.update(
         {
             "version": nereid.__version__,
@@ -26,6 +25,9 @@ class Settings(BaseSettings):
     VERSION: str = nereid.__version__
 
     FORCE_FOREGROUND: bool = False
+    ENABLE_ASYNC_ROUTES: bool = False
+    ASYNC_ROUTE_PREFIX: str = "/async"
+    ASYNC_MODE: Literal["none", "add", "replace"] = "none"
 
     # ALLOW_CORS_ORIGINS is a JSON-formatted list of origins
     # e.g: '["http://localhost", "http://localhost:4200", "http://localhost:3000", \
@@ -44,7 +46,7 @@ class Settings(BaseSettings):
     class Config:  # pragma: no cover
         env_prefix = "NEREID_"
         try:
-            with pkg_resources.path("nereid", ".env") as p:
+            with pkg_resources.path(".env") as p:
                 env_file = p
         except FileNotFoundError:
             pass
