+++ nereid/nereid/main.py
@@ -1,3 +1,3 @@
 from nereid.factory import create_app
 
-app = create_app(dict(NEREID_FORCE_FOREGROUND=True))
+app = create_app()
