+++ nereid/nereid/src/tasks.py
@@ -1,11 +1,11 @@
 from nereid.src.land_surface.tasks import land_surface_loading as land_surface_loading
+from nereid.src.network.tasks import network_subgraphs as network_subgraphs
 from nereid.src.network.tasks import (
-    network_subgraphs as network_subgraphs,
     render_solution_sequence_svg as render_solution_sequence_svg,
-    render_subgraph_svg as render_subgraph_svg,
-    solution_sequence as solution_sequence,
-    validate_network as validate_network,
 )
+from nereid.src.network.tasks import render_subgraph_svg as render_subgraph_svg
+from nereid.src.network.tasks import solution_sequence as solution_sequence
+from nereid.src.network.tasks import validate_network as validate_network
 from nereid.src.treatment_facility.tasks import (
     initialize_treatment_facilities as initialize_treatment_facilities,
 )
