+++ nereid/nereid/static/templates/display_svg.html
