+++ nereid/nereid/tests/test_api/test_watershed.py
@@ -35,6 +35,10 @@ def test_get_solve_watershed(client, watershed_responses, size, pct_tmnt):
     result_route = prjson.get("result_route")
 
     if result_route:
+        get_route = f"{settings.API_LATEST}/task/{prjson.get('task_id', 'error$!#*&^')}"
+        get_response = client.get(get_route)
+        assert get_response.status_code == 200, (prjson, get_route)
+
         get_response = client.get(result_route)
         assert get_response.status_code == 200
