+++ nereid/nereid/tests/test_src/test_land_surface/test_loading.py
@@ -89,7 +89,7 @@ def test_detailed_land_surface_loading_results(
     assert t["runoff_volume_cuft"].sum() > 0
 
     t = detailed_pollutant_loading_results(
-        land_surfaces_df,
+        t,
         wet_weather_parameters,
         dry_weather_parameters,
         seasons.keys(),
